Learnt how to amend the width of a class by a percentage instead of a pixel value. helps when the window is resized or another device is used.
Also noted that a class can be used alongside an element in the css file to specifically target that element within the scope of the tag being used in html.

Noted that a width of 50pc between two p elements will not be ideal unless you place the two p elements on the same line. This is due to a small space being added after each line.