Implemented the following:
html hr tag used for a horizontal rule. Initially this is a light grey line with the same color border.

The color can be amended with the background color property alongside the border color.

The height of the hr is denoted by the height property, initally set to 1px and 2x the border width - which is 1px.