CSS focus only:
Amended the footer anchor tag to be the color black by defualt.

Also utilised pseudo-selectors for events such as if-visited, if-hover, if-active aka whilst clicking. The colors have been amended for these too.

NOTE: it seems like best practice to align the css selectors with the order of the html file, so keep footer at the bottom if possible.