Learnt about how to use the HSL function instead of the rgb function.
HSL - hue saturation lightness
hue: 0-360 degrees
saturation: 0 to 100%
lightness: 0-100%

A saturation of 0 is a greyscaled image, whilst 100 is pure color.

lightness of 0 is pure black, 100 is pure white, 50pc is neutral.

Also learnt how to add linear gradients, given a direction in degrees and colours to utilise.
This is typically paired with the background property since it produces an image and the background property accepts this as an input.