Completed: Color stops and linear gradient function.
Color stops specifiy where a colors will transition along a linear gradient. They are denoted right after the color specified, with a space to seperate the two.

Linear gradient, specifies the direction in degrees from the vertical plane, alongside any number of colors... They could be an output of an rbg function, a hsl function or simply #hexcode.

The gradient direction is automatically 180deg (top to bottom) so this isnt required for some functions. Also if using 3 colors in a linear gradient, there is no point in specifying each color stop to be at 0, 50 and 100% respectively.