The labels are auto inline display elements, this can be amended in css with the display property.
Also learnt about a new unit measurement: REM... this stands for root em which is relative to the font size of the html element.

Inputs and associate labels added for name email and password... alongside a final submit input specified by a submit type and value.
The inputs and labels can be matched by the for and id attributes.

Utilised the email and password type inputs for each input respectively. They work the same as the text input however have additional features such as email validation and hiding text from the user as they type something in.

Learnt how to set boundaries on inputs such as with the minlength and pattern attributes. Typically pattern uses a regular expression to validify the users inputs.

Also implemented the required attribute for all inputs to ensure the user does not leave anything blank when trying to submit.