Final fieldset complete: profile pic... type file age... type number refferal... select - options bio... textarea
Learnt:
How to set a min and max number for age.
How to create a drop down selection of options.
Values to be left blank when an option is not useful to the server.
Amend the size of a textarea with the use of rows and cols attributes.

Also added unique names with the exception of the radio boxes which needed the same name in order to only accept one answer.