Added 9 photos in a gallary.
Learnt about default box sizing and how this affects image sizes based on the padding and border values.

Then amended this to suit the nature of the page. The better suited optioon here was border box > Which does the opposite to the default setting.
Images with a specified width STAY that width regardless of padding and border amendments.