Added calories and dividers
Again, rem is for root em > Relative to the font size of the html - I styled this earlier.

Amended calories info alignment with the use of margins to make sure everything fits together neatly.
Also ensured that it was set to a flex display with justification and alignment to ensure that the number appears to the right of the text.